fix for search becoming super hanged [NO-TEST][NO-BENCH]

+++ nina-chess/search_core.h
@@ -2,4 +2,4 @@
 #include "utils.h"
 
 inline constexpr uint32_t MAX_PLY = 512;
-inline constexpr uint32_t MAX_DEPTH = 256;
+inline constexpr int64_t MAX_DEPTH = 256;
